[#115 #116] refactor: 백앤드 업로드/로그인 부분 코드스멜 제거
